I think that Links stay around after they're closed, which means that you really want to set closing_ to false at some point.  Doing this in the open() routine seemed to make sense.
If you do NOT do this, then the second time a tcp convergence layer tries
to close out, it will NOT send a ContactDownEvent to the main daemon
(since link_->isclosing() is (still) true).  This in turn causes bundles
to not be sent over the link (unless you do a link close from the
command line, which seems to kick the thing).

There's probably a race condition in here somewhere, but this at least
makes dtnping run past the second bundle without hanging.