servlib/bundling/BundleDaemon.cc::event_handlers_completed: Correct sanity checking.
Code checks number of references remaining just before bundle is deleted.
OLD: Expected number of references to be 0.
NEW: Bundle will still be in ALL_BUNDLES at this point so expected number
of references will be 1. (Left over from introduction of ALL_BUNDLES?)
Improve warning message to tell user that it should be in ALL_BUNDLES
so that '1' is the right answer.