Move rg and ion_mode LTP convergence layer parameters to their rightful place.
These parameters were initially placed in the Link parameters structure.
Since they are LTP specific this is inappropriate.  The reason for this
was essentially that the 'link set_cl_defaults' was not implemented for
the LTP convergence layer (the method set_link_defaults was not
implemented).

This change implements this method and moves the parameters down into the
LTP convergence layer.  The opportunity is taken to improve the help text
for the various link configuration commands including showing only the
names of convergence layers compiled into the daemon (but showing all
possibilities). Suppliled by Elwyn Davies <davieseb@scss.tcd.ie>