Rejuvenates the DTN2 simulator which has been non-functional.
The main requirement is to ensure that the operations performed
work on the correct instance of the bundle daemon, particularly when
bundles are being transferred across a pseudo-link.  This was not handled
correctly in the existing code.  Additionally sim registrations were
converted to be derivatives of APIRegistrations rather than Registrations.
New registrations are now specified as EVENTSRC_APP rather than
EVENTSRC_ADMIN because of the restrictions on the regid of ADMIN
registrations to low numbers - less than the global store next_regid delivers!

The expiry time of registrations can now be specified in the configuration
file and registrations are correctly created.  Previously all
registrations appeared to expire in 10 secinds whcih was not enough
for many tests.  Sim registrations are also provided with a format
method and are specified as being derived from the Formatter class (first
on the super-class list).  This allows them to work with *%p but provokes
a compilation warning that the instance of Formatter in the super class is
inaccessible.  This is not a problem and the warning can be ignored.
Supplied by Elwyn Davies <davieseb@scss.tcd.ie>