ICN/BPQ patch set submitted by Elwyn Davies <davieseb@scss.tcd.ie>.
This set of changes, together with changeset 3551 is intended to support the
use of DTN2 with Information Centric Networking and the ni: URI scheme.  In
particular an extra application (apps/dtnpublish) has been written to allow
bundles containing a Query Extension (BPQ) block of the (new) kind KIND_PUBLISH
to be sent out into the network where they will be cached by appropriately
configured nodes to which they are forwarded.  The following changes have been
made:

- a new application 'apps/dtnpublish' has been written.  It has numerous
options - see the dtnpublish manual page or use 'dtnpublish --help'.

- Bundles that are in the BPQ cache will not be subject to early deletion so
they will remain avaioable in the cache until their lifetime expires or cache
space runs out so they have to be deleted as the least recently used.

- Bundles of KIND_PUBLISH *are* cached in the node where the application
API created the bundle with a Query Extension Block (as opposed to responses).
The KIND_RESPONSE bundles were previoulsy not cached because of the bug
mentioned below which prevented the response bundles being returned correctly.
This problem is now fixed, but at present KIND_RESPONSE blocks are NOT cached
in the node where the bundle is originated.   This is partly because there is a
logical problem if the response was externally generated from an existing
cached bundle.  This situation may be altered in future.

- Responses derived from cached bundles that were cached with KIND_PUBLISH
BPQBlocks, will have their responses mutated to KIND_RESPONSE.

In the process a number of bugs identified with the initial Query Extension
block (BPQ) code (changeset 3528):

- Bundles have to be deleted from the appropriate BPQ cache entry when they
expire or are otherwise deleted.  If the cache entry is then empty it has to be
deleted also.

- An apparently long standing bug was identified and fixed: bundles containing
extension blocks and metadata blocks that were received from the API and
delivered directly to an application on the same node would not have the
extension blocks or metadata blocks attached to the delivered bundle.  This was
because APIServer::dtn_recv method did not scan either the api_blocks or
generated_metadata vectors when creating the structure to pass back to the
application.

- The BLockInfo::type method now allows for all the different sorts of
extension block that have been defined.

Some general improvements to the BPQ-related code have been made:

- To avoid repeated recreation of the internal representation of the data in the
Query Extension Block, a BPQBlock instance is created when the BlockInfo
instance representing the Query Extension Block is created :
    * The dtn_send API function is called with a BPQ extension block
    * A bundle is received form a link containing a BPQ extension block
    * A bundle is read back from persistent storage on bundle daemon startup
    * A response bundle is created as a result of receiving a BPQ QUERY

Note that it is not necessary to duplicate the BPQBlock into BlockInfo instances
placed into xmit_blocks vectors created during bundle transmission as all the
routines processing these blocks treat them as opaque data.
To manage this data BPQBlock has been made a derived class of BP_Local.  The
BlockInfo class provides for the storage of a pointer to a BP_Local class
instance which can be written and accessed by set_locals() and locals() methods
respectively. A BPQBlock instance is created from the opaque data in the
extension block when the BlockInfo is created.  Classes that are designed to
process BPQ blocks can then 'downcast' the BP_Local pointer type to the BPQBlock
pointertyoe using dynamic_cast.  This capability existed in DTN2 but was not
previously used for other sorts of extension blocks.

- As part of the above change, the original creation timestamp and source EID
have to be written into the opaque data (and the BPQBlock) when a bundle with a
Query Extension Block is received in the APIServer::dtn_send method.  This has
been implemented by adding an extra Bundle pointer parameter to the
BlockProcessor::init_block method and making this method virtual so it can be
overloaded in other types of BlockProcessor.  The extra paramater is ignored by
the generic method in the base class (BlockProcessor) but, if it is not NULL,
is used by the overridden method in BPQBlockProcessor to set the creation
timestamp and original source EID.  In the other cases where new BlockInfo
structures are created for Query Extension blocks, the opaque data used will
already contain the correct timestamp and source EID so the Bundle pointer
parameter is left as NULL which signals the init_block routine to leave the
opaque data unchanged.

- In association with this the initialization methods for BPQBlock have been
restructured so that a BPQBlock instance can be created directly from the
opaque data in the BlockInfo instance rather than requiring the data in the
bundle each time.  It is also nor longer necessary to differentiate between
locally and remotely created BPQBlocks.

- Similarly, a specialized BPQBlockProcessor::reload_post_process method has
been added to create the BPQBlock instance and save it in the BlockInfo
instance when pulling bundles in from persistent storage.

- The BPQBlockProcessor::consume method now creates the BPQBlock and records
it in the BlockInfo instance when the block is complete.

- New setters have been added to update the creation_ts and original_source
fields in BPQBlock and a validated flag has been added to record if the block
is correctly populated.

- BPQBlockProcessor::generate has been modified to make better use of the
'source' block set up by BPQBlockProcessor::prepare rather than creating
information afresh.

To assist with management and monitoring, some additions and improvements have
been made to the BPQ commands and the bundle info command:

- added command 'bpq list' which lists the query values for the current BPQ
cache entries

- added command 'bpq lru' which lists the query values in the BPQ cache in the
order of the least recently used (LRU) list with the most recently used item
first.

- improved 'bundle info <bundleid>' command so that it prints the contents of
the BPQ block as well as its existence and size.  THis is done by provdiding
an overloaded 'BPQBlockProcessor::format' method rather than using the base
class version.

- Additional locks have been provided to ensure that the whole BPQ cache is
locked when manipulating the list of cacche entries as well as the existing
locks for each fragment list.