DTN2 IPND Reference Implementation ==================================
This version of DTN2 includes a reference implementation of DTN IP Neighbor
Discovery (IPND) as described in the IRTF Internet Draft
(draft-irtf-dtnrg-ipnd-02). This implementation is a feature separate from the
IP Discovery functionality that is included by default in DTN2, and must be
enabled as an experimental feature by configure. Provided by
Angela Hennessy <ahennes1@math.umd.edu>