Ensure all convergence layers are consistent with persistent link interface.
Changeset 3544 added the infrastructure needed to allow links to be made
persistent and support the persistent storing of the forwarding log for bundles.
It appears that the update was not fully tested with all the various convergence
layers.  Sergey Sireskin reported that the NORM convergence layer no longer
compiles correctly as of 2012/12/07.

This changeset updates all the active convergence layers (not including the File
CL) to be consistent and fully integrated with the persistent link changes.

The changes involve:
- Ensuring that every externally usable CL (some CL classes are used as base
  classes and cannot be used as actual CLs) has an init_link method.
- Ensuring that every CL has a parse_link_params method, that these methods
  deal with the parameters specific to the class and call the parent class
  method to deal with any parameters specific to a parent class (possibly
  recursively).
- Ensuring that every externally usable CL has a set_link_defaults method that
  configures the link parameters to be used when a link is created
  opportunistically using this CL.  This method is called by the
  'link set_link_defaults' command in the control interface.
- Ensuring that every CL has a dump_link method to print out the parameter
  values.

The Convergence Layer classes affected are:
- AX25
- Bluetooth
- Connection
- Ethernet
- NORM
- Null
- SrqPacket
- Serial
- TCP
- UDP

Unrelated item:
Adds a debugging log statement to record whether conventional or BBN IPND IP
discovery is active.

Elwyn Davies @ 2013/01/10