This patch fixes proactive fragmentation and fragment reassembly when extension blocks are present.
It guesses the amount of payload to include in each fragment by
assuming that the extension blocks in the fragment will be the
same size as the extension blocks in the original bundle. Any link
can have an mtu set from the console. It additionally assumes
that locally generated blocks in the original
bundle will be generated (with the same size) in each fragment.
This is all painfully kludgy. The saving grace is that the worst
case is that we have to fragment multiple times, because the
fragments actually end up larger than the MTU.

Patch provided by Amy Alford <aloomis@sarn.org>